[OMNIML-2244] Create the nvfp4 quant exporter
