+++ examples/onnx_ptq/llm_export.py
@@ -30,7 +30,7 @@ from packaging.version import Version
 from transformers import AutoConfig, AutoTokenizer
 
 import modelopt
-from modelopt.onnx.export import INT4QuantExporter
+from modelopt.onnx.export import INT4QuantExporter, NVFP4QuantExporter
 from modelopt.onnx.llm_export_utils.export_utils import (
     ModelLoader,
     WrapperModelForCausalLM,
@@ -38,7 +38,6 @@ from modelopt.onnx.llm_export_utils.export_utils import (
 )
 from modelopt.onnx.llm_export_utils.quantization_utils import quantize
 from modelopt.onnx.llm_export_utils.surgeon_utils import fold_fp8_qdq_to_dq
-from modelopt.onnx.quantization.qdq_utils import fp4qdq_to_2dq
 from modelopt.torch.export import export_hf_checkpoint
 from modelopt.torch.quantization.utils import is_quantized_linear
 
@@ -275,7 +274,7 @@ def surgeon_llm(
 
     if dtype == "nvfp4":
         with time_operation("quantizing weights to nvfp4"):
-            onnx_model = fp4qdq_to_2dq(onnx_model, verbose=True)
+            onnx_model = NVFP4QuantExporter.process_model(onnx_model)
 
     elif dtype == "int4_awq":
         with time_operation("quantizing weights to int4"):
