+++ modelopt/onnx/quantization/qdq_utils.py
@@ -24,7 +24,6 @@ import onnx_graphsurgeon as gs
 import torch
 from onnx import numpy_helper
 
-from modelopt.onnx import utils
 from modelopt.onnx.logging_config import logger
 from modelopt.onnx.quantization.graph_utils import (
     get_tensor_consumer_nodes,
@@ -32,16 +31,8 @@ from modelopt.onnx.quantization.graph_utils import (
     get_tensor_producer_nodes,
     remove_redundant_cast_nodes,
 )
-from modelopt.onnx.quantization.quant_utils import (
-    compute_e8m0,
-    get_amax,
-    get_num_bits,
-    get_weights_scaling_factor,
-    get_weights_scaling_factor_2,
-    quantize,
-)
+from modelopt.onnx.quantization.quant_utils import compute_e8m0, get_amax, get_num_bits
 from modelopt.onnx.utils import get_attribute, has_attribute
-from modelopt.torch.quantization.qtensor import NVFP4QTensor
 
 QUANTIZE_NODE_NAME = "QuantizeLinear"
 DEQUANTIZE_NODE_NAME = "DequantizeLinear"
@@ -692,25 +683,6 @@ def _cast_fp8(array: np.ndarray) -> np.ndarray:
     return array_f8
 
 
-def _cast_fp4(array: np.ndarray) -> np.ndarray:
-    """Cast a numpy array to FLOAT4E2M1 using PyTorch.
-
-    Note: The first dimension of the array must be divisible by 2
-    as two FP4 values are packed into a single byte.
-    """
-    array_f32_t = torch.from_numpy(array)
-    array_f32_t_shape = array_f32_t.shape
-    assert array_f32_t_shape[0] % 2 == 0, "array_f32_t_shape[0] must be divisible by 2"
-    array_f4_t_shape = (array_f32_t_shape[0] // 2, *array_f32_t_shape[1:])
-    if torch.cuda.is_available():
-        array_f32_t = array_f32_t.cuda()
-    array_f4_t = NVFP4QTensor._cast_fp4(array_f32_t)
-    array_f4_t = array_f4_t.flatten()
-    array_f4_t_packed = (array_f4_t[::2] | (array_f4_t[1::2] << 4)).reshape(array_f4_t_shape)
-    array_f4 = array_f4_t_packed.cpu().numpy().astype(np.uint8)
-    return array_f4
-
-
 def _create_fp8_tensor(scaled: np.ndarray, weight_name: str) -> onnx.TensorProto:
     """Create a FLOAT8E4M3FN tensor directly from numpy array."""
     fp8_data = _cast_fp8(scaled)
@@ -1160,237 +1132,5 @@ def quantize_weights_to_mxfp8(
             if attr.name == "approximate":
                 attr.s = b"tanh"
                 logger.debug(f"Updated GELU node {node.name} to use tanh approximation")
-
-    return onnx_model
-
-
-def replace_fp4qdq_with_2dq(
-    graph: onnx.GraphProto,
-    node: onnx.NodeProto,
-    initializer_indices: dict[str, int],
-    value_info_map: dict[str, onnx.ValueInfoProto],
-    graph_inputs: set[str],
-    w_f4: np.ndarray,
-    sw_f32_per_tensor: np.ndarray,
-    sw_f8_per_block: np.ndarray,
-    block_size: int,
-):
-    """Replaces the given node in the ONNX graph with a subgraph consisting of two DequantizeLinear nodes.
-
-    Args:
-        graph: The ONNX graph containing the node to replace.
-        node: The node to be replaced.
-        initializer_indices: A dictionary mapping initializer names to their indices in the graph.
-        value_info_map: A dictionary mapping value info names to their ValueInfoProto objects.
-        graph_inputs: A set of graph input names.
-        w_f4: NumPy array for w_f4.
-        sw_f32_per_tensor: NumPy array for sw_f32_per_tensor.
-        sw_f8_per_block: NumPy array for sw_f8_per_block.
-        block_size: Block size used in block quantization.
-    """
-
-    def _add_initializer(initializer):
-        if initializer.name not in initializer_indices:
-            graph.initializer.append(initializer)
-
-    def _add_input_value_info(graph, tensor_proto):
-        assert tensor_proto.name not in graph_inputs, (
-            f"{tensor_proto.name} already in graph inputs."
-        )
-        assert tensor_proto.name not in value_info_map, (
-            f"{tensor_proto.name} already in value info."
-        )
-
-        value_info = onnx.helper.make_tensor_value_info(
-            tensor_proto.name, tensor_proto.data_type, tensor_proto.dims
-        )
-        graph.input.append(value_info)
-
-    # Remove the original node from the graph
-    graph.node.remove(node)
-    weight_name = node.input[0]
-
-    # Generate unique names for the initializers
-    w_f4_name = weight_name + "_f4"
-    sw_f8_per_block_name = weight_name + "_f8_scale"
-    sw_f32_per_tensor_name = sw_f8_per_block_name + "_f32_scale"
-
-    # Create TensorProto for initializers
-    w_f4_proto = onnx.helper.make_tensor(
-        name=w_f4_name,
-        data_type=onnx_dtype_map["Float4"],
-        dims=[w_f4.shape[0] * 2, *w_f4.shape[1:]],
-        vals=w_f4.tobytes(),
-        raw=True,
-    )
-    sw_f32_per_tensor_proto = onnx.numpy_helper.from_array(
-        sw_f32_per_tensor, sw_f32_per_tensor_name
-    )
-    sw_f8_per_block_proto = onnx.numpy_helper.from_array(sw_f8_per_block, sw_f8_per_block_name)
-    sw_f8_per_block_proto = onnx.helper.make_tensor(
-        name=sw_f8_per_block_name,
-        data_type=onnx_dtype_map["Float8"],
-        dims=[*sw_f8_per_block.shape],
-        vals=sw_f8_per_block.tobytes(),
-        raw=True,
-    )
-
-    # Add ValueInfo for the initializers if not present
-    _add_input_value_info(graph, w_f4_proto)
-    _add_input_value_info(graph, sw_f32_per_tensor_proto)
-    _add_input_value_info(graph, sw_f8_per_block_proto)
-
-    # Add the initializers to the graph
-    _add_initializer(w_f4_proto)
-    _add_initializer(sw_f32_per_tensor_proto)
-    _add_initializer(sw_f8_per_block_proto)
-
-    # Create DequantizeLinear_1 node: (sw_f8_per_block, sw_f32_per_tensor) -> sw_f32
-    sw_f32_name = weight_name + "_f32_scale"
-    dequant1 = onnx.helper.make_node(
-        "DequantizeLinear",
-        inputs=[sw_f8_per_block_proto.name, sw_f32_per_tensor_proto.name],
-        outputs=[sw_f32_name],
-        name=weight_name + "_DequantizeLinear",
-    )
-
-    # Create DequantizeLinear_2 node: (w_f4, sw_f32) -> w_32
-    w32_name = node.output[0]
-    dequant2 = onnx.helper.make_node(
-        "DequantizeLinear",
-        inputs=[w_f4_proto.name, sw_f32_name],
-        outputs=[w32_name],
-        name=weight_name + "_DequantizeLinear_1",
-        axis=-1,
-        block_size=block_size,
-    )
-
-    # Add value_info for sw_f32
-    # Assuming sw_f16 has the same shape as sw_f8_per_block
-    sw_f32_type_proto = onnx.helper.make_tensor_type_proto(
-        elem_type=onnx_dtype_map["Float"], shape=sw_f8_per_block.shape
-    )
-    sw_f16_value_info = onnx.helper.make_value_info(name=sw_f32_name, type_proto=sw_f32_type_proto)
-    graph.value_info.append(sw_f16_value_info)
-
-    # Change the data type of w16 (output of 2nd DQ) to model weight precision type
-    if w32_name in value_info_map:
-        value_info_map[w32_name].type.tensor_type.elem_type = onnx_dtype_map["Float"]
-    else:
-        raise ValueError(f"ValueInfo for {w32_name} not found.")
-
-    # Add the new nodes to the graph
-    graph.node.extend([dequant1, dequant2])
-
-
-def fp4qdq_to_2dq(onnx_model: onnx.ModelProto, verbose: bool = False) -> onnx.ModelProto:
-    """Convert FP32/FP16 weights of the given ONNX model to FP4 weights and scaling factors.
-
-    TRT_FP4QDQ nodes will get removed from the weights and have two DQ nodes with those converted FP4
-    weights and scaling factors in the output model.
-
-    Args:
-        onnx_model: ONNX model protobuf.
-
-    Returns:
-        ONNX model protobuf with DQ nodes for weights and DynQ + DQ nodes for activations.
-    """
-    logger.info("Converting model with FP4QDQ nodes to 2DQ only model")
-    graph = onnx_model.graph
-    initializers = graph.initializer
-    initializers_to_delete = []
-    tensor_consumers = get_tensor_consumer_nodes(graph)
-    initializer_indices = {
-        initializer.name: idx for idx, initializer in enumerate(graph.initializer)
-    }
-    value_info_map = {vi.name: vi for vi in graph.value_info}
-    graph_inputs = {inp.name for inp in graph.input}
-
-    def _cast_input_dtypes(node: onnx.NodeProto, precision_dtype: str):
-        # Change the input types to match weight precision (precision_dtype)
-        if node.op_type == "Transpose":
-            maybe_matmul = tensor_consumers[node.output[0]][0]
-            assert maybe_matmul.op_type == "MatMul"
-            node = maybe_matmul
-
-        # Create Cast nodes for each input of the target node except bias
-        for i, input_name in enumerate(node.input[:2]):
-            cast_output_name = input_name + "_f16"  # Unique name for the cast output
-
-            # Create a Cast node to convert the input to FP16/BF16
-            cast_node = onnx.helper.make_node(
-                "Cast",
-                inputs=[input_name],  # Original input of the target node
-                outputs=[cast_output_name],
-                to=onnx_dtype_map[precision_dtype],  # Cast to FP16/BF16
-            )
-
-            # Insert the Cast node into the graph
-            graph.node.extend([cast_node])
-
-            # Update the target node input to use the cast node output
-            node.input[i] = cast_output_name
-
-    def _get_precision_dtype() -> str:
-        # Check initializers to determine the precision of the weights
-        precision_dtype = "Half"
-        for initializer in graph.initializer:
-            if initializer.data_type == 16:
-                precision_dtype = "BFloat16"
-                break  # Assuming all weights are of the same precision
-
-        return precision_dtype
-
-    if verbose:
-        logger.info("Post-processing TRT_FP4QDQ nodes for TRT deployment")
-    precision_dtype = _get_precision_dtype()
-    logger.debug(f"Using precision dtype: {precision_dtype}")
-    fp4_qdq_nodes = [node for node in graph.node if node.op_type == "TRT_FP4QDQ"]
-    logger.debug(f"Found {len(fp4_qdq_nodes)} FP4QDQ nodes to convert")
-
-    for node in fp4_qdq_nodes:
-        idx1 = initializer_indices.get(node.input[0], None)
-        assert idx1 is not None, f"Initializer for weight '{node.input[0]}' not found."
-        block_size = node.attribute[0].i
-        initializers_to_delete.append(initializers[idx1].name)
-        logger.debug(
-            f"Processing FP4QDQ node for weight {node.input[0]} with block size {block_size}"
-        )
-
-        tensor = initializers[idx1]
-        w32 = utils.read_f16_tensor_as_fp32(tensor)
-        sw_f32_per_tensor = get_weights_scaling_factor_2(w32)
-        sw_f32_per_block = get_weights_scaling_factor(w32, block_size, sw_f32_per_tensor)
-        w_f32 = quantize(w32, block_size, sw_f32_per_block, sw_f32_per_tensor)
-
-        # Real quantize the tensors
-        w_f4 = _cast_fp4(w_f32)
-        sw_f8_per_block = _cast_fp8(sw_f32_per_block)
-
-        replace_fp4qdq_with_2dq(
-            graph,
-            node,
-            initializer_indices,
-            value_info_map,
-            graph_inputs,
-            w_f4,
-            sw_f32_per_tensor,
-            sw_f8_per_block,
-            block_size,
-        )
-
-        # We need to change the bias etc. type
-        next_node = tensor_consumers[node.output[0]][0]
-        _cast_input_dtypes(next_node, precision_dtype)
-
-        if verbose:
-            logger.debug(f"Replaced {node.name} with 2 DQ nodes")
-
-    new_initializers = [
-        init for init in graph.initializer if init.name not in initializers_to_delete
-    ]
-    graph.ClearField("initializer")
-    graph.initializer.extend(new_initializers)
-    logger.info(f"Removed {len(initializers_to_delete)} initializers")
 
     return onnx_model
